Move escape handler to modal
